+++ installer/templates/phx_umbrella/apps/app_name_web/config/prod.secret.exs
@@ -12,7 +12,7 @@ config :<%= web_app_name %>, <%= endpoint_module %>,
   ],
   secret_key_base: secret_key_base
 
-# ## Using releases (Elixir v1.9+)
+# ## Using releases
 #
 # If you are doing OTP releases, you need to instruct Phoenix
 # to start each relevant endpoint: