+++ test/mix/phoenix_test.exs
@@ -76,12 +76,12 @@ defmodule Mix.PhoenixTest do
       admin: true,
       meta: %{},
       name: "some name",
-      date_of_birth: %Date{year: 2010, month: 4, day: 17},
-      happy_hour: %Time{hour: 14, minute: 0, second: 0},
-      happy_hour_usec: %Time{hour: 14, minute: 0, second: 0, microsecond: {0, 6}},
-      joined: %NaiveDateTime{year: 2010, month: 4, day: 17, hour: 14, minute: 0, second: 0},
-      joined_utc: "2010-04-17T14:00:00Z",
-      joined_utc_usec: "2010-04-17T14:00:00.000000Z",
+      date_of_birth: ~D[2010-04-17],
+      happy_hour: ~T[14:00:00],
+      happy_hour_usec: ~T[14:00:00.000000],
+      joined: ~N[2010-04-17 14:00:00],
+      joined_utc: ~U[2010-04-17 14:00:00Z],
+      joined_utc_usec: ~U[2010-04-17 14:00:00.000000Z],
       token: "7488a646-e31f-11e4-aace-600308960662",
       other: "some other"
     }